model config component and integ in calibrate
